requirements.txt
tweepy==4.14.0
python-dotenv==1.0.0
requests==2.31.0
vaderSentiment==3.3.2
textblob==0.18.0.post0